+++ src/Commands/MakeEndpointCommand.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Anteris\Autotask\Generator\Command;
+namespace Anteris\Autotask\Generator\Commands;
 
-use Anteris\Autotask\Generator\Helper\Api;
 use Illuminate\Support\Env;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class MakeDefaultsCommand extends AbstractMakeCommand
+class MakeEndpointCommand extends AbstractMakeCommand
 {
     /** @var string The name of this command.  */
-    protected static $defaultName = 'make:defaults';
+    protected static $defaultName = 'make:endpoint';
 
     /**
      * Configures the command information.
@@ -21,11 +21,11 @@ class MakeDefaultsCommand extends AbstractMakeCommand
     protected function configure()
     {
         $this
-            ->setDescription('Creates a new PHP class for the default Autotask services.')
+            ->setDescription('Creates a new PHP class for interacting with an Autotask service.')
+            ->addArgument('entity', InputArgument::REQUIRED, 'The entity to generate classes for.')
             ->addOption('no-cache', null, InputOption::VALUE_NONE, 'Specifies whether any previous cached files should be ignored (still makes use of cache for efficiency, but clears afterward).')
-            ->addOption('force', 'f', InputOption::VALUE_NONE, 'Overwrites previous classes if they exist.')
-            ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Sets the output directory for the generated classes.');
-
+            ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Sets the output directory for the generated class.')
+            ->addOption('force', 'f', InputOption::VALUE_NONE, 'Overwrites previous classes if they exist.');
     }
 
     /**
@@ -34,13 +34,10 @@ class MakeDefaultsCommand extends AbstractMakeCommand
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->setupEnvironment($input, $output);
-        $endpoints = Api::endpoints();
 
         try {
-            foreach ($endpoints as $service) {
-                $output->writeln('Generating classes for ' . $service);
-                $this->generator->makeResource($service);
-            }
+            $output->writeln('Generating classes for ' . $input->getArgument('entity'));
+            $this->generator->makeResource($input->getArgument('entity'));
         } catch (\Exception $error) {
             $output->writeln(
                 '<error>There was an error creating that endpoint: ' .
@@ -51,10 +48,8 @@ class MakeDefaultsCommand extends AbstractMakeCommand
         }
 
         $output->writeln(
-            '<info>Successfully created '.
-            count($endpoints).
-            ' endpoints in '.
-            (microtime(true) - $_SERVER['REQUEST_TIME_FLOAT']).
+            '<info>Successfully created endpoint in ' .
+            (microtime(true) - $_SERVER['REQUEST_TIME_FLOAT']) .
             ' seconds!</info>'
         );
         $output->writeln('');